Move projectInternalView to a JVM-specific interface
