+++ subprojects/core-api/src/main/java/org/gradle/api/artifacts/dsl/Dependencies.java
@@ -20,8 +20,6 @@ import org.gradle.api.Incubating;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.ExternalModuleDependency;
 import org.gradle.api.artifacts.ProjectDependency;
-import org.gradle.api.attributes.CompileView;
-import org.gradle.api.model.ObjectFactory;
 
 import javax.annotation.Nullable;
 import javax.inject.Inject;
@@ -29,7 +27,8 @@ import javax.inject.Inject;
 /**
  * Universal APIs that are available for all {@code dependencies} blocks.
  *
- * @apiNote This API is <strong>incubating</strong> and is likely to change until it's made stable.
+ * @apiNote This API is likely to change until it's made stable.
+ *
  * @implSpec These methods are not intended to be implemented by end users or plugin authors.
  * @since 7.6
  */
@@ -44,14 +43,6 @@ public interface Dependencies {
      */
     @Inject
     DependencyFactory getDependencyFactory();
-
-    /**
-     * Injected service to create named objects.
-     *
-     * @return injected service
-     */
-    @Inject
-    ObjectFactory getObjectFactory();
 
     /**
      * The current project. You need to use {@link #project()} or {@link #project(String)} to add a {@link ProjectDependency}.
@@ -84,23 +75,6 @@ public interface Dependencies {
      */
     default ProjectDependency project() {
         return getDependencyFactory().create(getProject());
-    }
-
-    /**
-     * Create a dependency on the current project's internal view. During compile-time, this dependency will
-     * resolve the current project's implementation in addition to its API. During runtime, this dependency
-     * behaves as a usual project dependency.
-     *
-     * @return the current project, including implementation details, as a dependency
-     *
-     * @since 8.0
-     */
-    default ProjectDependency projectInternalView() {
-        ProjectDependency currentProject = project();
-        currentProject.attributes(attrs -> {
-            attrs.attribute(CompileView.VIEW_ATTRIBUTE, getObjectFactory().named(CompileView.class, CompileView.JAVA_INTERNAL));
-        });
-        return currentProject;
     }
 
     /**
